requirements.txt
# Core Framework
scrapy==2.11.0
scrapy-redis==0.7.3

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
asyncpg==0.29.0
psycopg[binary,pool]==3.1.18
redis[hiredis]==5.0.1

# Message Queue
aiokafka==0.10.0

# Storage
miniopy-async==1.19

# Scheduler
apscheduler==4.0.0a5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# Monitoring
prometheus-client==0.19.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Data Processing
lxml==5.1.0
parsel==1.8.1

# Utilities
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
